Rename some files and fix namespaces
