+++ Events/Actions/HealAction.cs
@@ -2,17 +2,20 @@
 using Microsoft.Xna.Framework;
 using TurnBasedFeest.Actors;
 using TurnBasedFeest.BattleSystem;
-using Microsoft.Xna.Framework.Graphics;
 using TurnBasedFeest.Utilities;
+using Microsoft.Xna.Framework.Graphics;
 
-namespace TurnBasedFeest.Actions
+namespace TurnBasedFeest.Events.Actions
 {
-    class ActionDefend : IAction
+    class HealAction : IAction
     {
         int actionTime = 1000;
         int elapsedTime;
         Actor source;
         Actor target;
+        int beginHP;
+        int targetHP;
+        int heal = 20;
 
         public void SetActors(Actor source, Actor target)
         {
@@ -23,16 +26,18 @@ namespace TurnBasedFeest.Actions
         public void Initialize()
         {
             elapsedTime = 0;
-            this.source.health.color = Color.Violet;
+            beginHP = source.health.actorCurrentHealth;
+            targetHP = (source.health.actorCurrentHealth + heal >= source.health.actorMaxHealth) ? source.health.actorMaxHealth : (source.health.actorCurrentHealth + heal);
         }
 
         public bool Update(Battle battle, Input input)
         {
-            elapsedTime += (int) Game1.time.ElapsedGameTime.TotalMilliseconds;
-            
-            if(elapsedTime > actionTime)
+            elapsedTime += (int)Game1.time.ElapsedGameTime.TotalMilliseconds;
+
+            source.health.actorCurrentHealth = (elapsedTime >= actionTime) ? targetHP : (int) MathHelper.SmoothStep(beginHP, targetHP, (elapsedTime / (float)actionTime));
+
+            if (source.health.actorCurrentHealth == targetHP)
             {
-                source.health.color = Color.White;
                 battle.currentActor.battleEvents.RemoveAt(battle.eventIndex);
                 battle.eventIndex--;
                 return true;
@@ -40,12 +45,13 @@ namespace TurnBasedFeest.Actions
             else
             {
                 return false;
-            }            
+            }
+            
         }
 
         public string GetName()
         {
-            return "Defend";
+            return "Heal";
         }
 
         public void Draw(Battle battle, SpriteBatch spritebatch, SpriteFont font)
